Fixes #22723: Backport scala 2.13 change for scala 3.0 migration
