+++ webapp/sources/ldap-inventory/inventory-api/src/main/scala/com/normation/inventory/domain/DataTypes.scala
@@ -196,9 +196,11 @@ final case class Certificate(value: String) extends SecurityToken {
  * Comparison are really important in Version
  */
 final class Version(val value: String) extends AnyVal {
-  override def toString()                = "[%s]".format(value)
+  override def toString() = "[%s]".format(value)
 }
-given Ordering[Version] = _.value compareTo _.value
+object Version {
+  implicit val ord: Ordering[Version] = _.value compareTo _.value
+}
 
 object InventoryProcessingLogger extends NamedZioLogger {
   override def loggerName: String = "inventory-processing"
