+++ webapp/sources/ldap-inventory/inventory-api/src/main/scala/com/normation/inventory/domain/Identifiers.scala
@@ -47,7 +47,7 @@ trait Uuid extends Any {
 }
 
 // TODO: migration-scala3 - bug: https://github.com/lampepfl/dotty/issues/16467
-final case class NodeId(val value: String) extends /*AnyVal with*/ Uuid
+final case class NodeId(val value: String) extends AnyVal with Uuid
 
 final case class MachineUuid(val value: String) extends AnyVal with Uuid
