+++ webapp/sources/ldap-inventory/inventory-api/src/main/scala/com/normation/inventory/domain/MemorySize.scala
@@ -51,9 +51,11 @@ final case class MemorySize(size: Long) extends AnyVal {
     value + " " + unit
   }
 }
-given Ordering[MemorySize] = _.size compareTo _.size
 
 object MemorySize {
+
+  val ord: Ordering[MemorySize] = _.size compareTo _.size
+
   /*
    * We should accept:
    * - no decimal numbers
