+++ webapp/sources/ldap-inventory/inventory-repository/pom.xml
@@ -58,11 +58,11 @@ along with Rudder.  If not, see <http://www.gnu.org/licenses/>.
 
     <dependency>
       <groupId>net.liftweb</groupId>
-      <artifactId>lift-json_${scala2-binary-version}</artifactId>
+      <artifactId>lift-json_${scala-binary-version}</artifactId>
     </dependency>
     <dependency>
       <groupId>net.liftweb</groupId>
-      <artifactId>lift-util_${scala2-binary-version}</artifactId>
+      <artifactId>lift-util_${scala-binary-version}</artifactId>
     </dependency>
   </dependencies>
