+++ webapp/sources/rudder/rudder-core/src/main/scala/com/normation/rudder/services/policies/DeploymentService.scala
@@ -1817,7 +1817,7 @@ object RuleExpectedReportBuilder extends Loggable {
                   case None     =>
                     varReportId match {
                       case None     =>
-                        innerExpandedVars.zip(innerUnexpandedVars).map(ExpectedValueMatch.apply)
+                        innerExpandedVars.zip(innerUnexpandedVars).map(ExpectedValueMatch.tupled)
                       case Some(id) =>
                         innerExpandedVars.map(v => ExpectedValueId(v, id))
                     }
@@ -1862,7 +1862,7 @@ object RuleExpectedReportBuilder extends Loggable {
       List(
         ValueExpectedReport(
           technique.id.name.value,
-          trackingVarCard._1.toList.zip(trackingVarCard._2.toList).map(ExpectedValueMatch.apply)
+          trackingVarCard._1.toList.zip(trackingVarCard._2.toList).map(ExpectedValueMatch.tupled)
         )
       )
     } else {
