+++ webapp/sources/rudder/rudder-web/src/test/scala/bootstrap/liftweb/RudderUserDetailsTest.scala
@@ -40,11 +40,13 @@ package bootstrap.liftweb
 import org.junit.runner.RunWith
 import org.specs2.mutable._
 import org.specs2.runner.JUnitRunner
+import scala.annotation.nowarn
 
 /*
  * Test hash algo for user password.
  */
 
+@nowarn("msg=a type was inferred to be `AnyVal`")
 @RunWith(classOf[JUnitRunner])
 class RudderUserDetailsTest extends Specification {
